Move RuntimeShape into its own file and make it smaller on TFLM.
keyword_benchmark size before:
   text    data     bss     dec
  57820   37671   25612  121103

keyword_benchmark size after:
   text    data     bss     dec
  56860   37671   25612  120143
